requirements.txt
sanic
mysql-connector-python
faker
passlib
sanic-ext
